Improved handling of major/minor updates
Previously, the `major` property for a build in the manifest file could
only be used to force a build to major regardless of the version number,
and so we always left it as `false`. This change causes the first three
characters of the version number to be used (e.g., 7.0 > 6.1) as long as
`major` is `null` and otherwise allows the build type to be forced with
`true` or `false` (e.g., to eventually set a major upgrade as minor so
that it starts auto-upgrading instead of prompting).